Move appConfig.js to src director and change relative file paths
