+++ src/app/components/BibPage/SubjectHeadings.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router';
-import appConfig from '../../../../appConfig';
+import appConfig from '../../appConfig';
 
 const constructSubjectHeading = (heading, i) => {
   const { uuid, parent, label } = heading
