+++ src/app/components/Breadcrumbs/Breadcrumbs.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Link } from 'react-router';
 
 import { trackDiscovery } from '../../utils/utils';
-import appConfig from '../../../../appConfig';
+import appConfig from '../../appConfig';
 
 const {
   baseUrl,
