+++ src/app/components/Results/ResultsList.jsx
@@ -13,7 +13,7 @@ import {
   trackDiscovery,
 } from '../../utils/utils';
 import ItemTable from '../Item/ItemTable';
-import appConfig from '../../../../appConfig';
+import appConfig from '../../appConfig';
 
 class ResultsList extends React.Component {
   constructor(props) {
